requirements.txt
amqp==5.2.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
Django==4.2.8
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
kombu==5.3.4
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-celery==0.0.0
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
